requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
motor==3.3.2
python-dotenv==1.0.1
pydantic==1.10.13
pymongo==4.5.0